requirements.txt
requests==2.32.5
python-dotenv==1.2.1
isort==7.0.0